Renames library from `baselines` to `ocpmodels`
