+++ ocpmodels/datasets/iso17.py
@@ -1,7 +1,7 @@
 from torch_geometric.data import DataLoader
 
-from baselines.common.registry import registry
-from baselines.datasets.base import BaseDataset
+from ocpmodels.common.registry import registry
+from ocpmodels.datasets.base import BaseDataset
 
 
 @registry.register_dataset("iso17")
