+++ ocpmodels/datasets/qm9.py
@@ -2,9 +2,9 @@ import torch_geometric.transforms as T
 from torch_geometric.data import DataLoader
 from torch_geometric.datasets import QM9
 
-from baselines.common.registry import registry
-from baselines.common.utils import Complete
-from baselines.datasets.base import BaseDataset
+from ocpmodels.common.registry import registry
+from ocpmodels.common.utils import Complete
+from ocpmodels.datasets.base import BaseDataset
 
 
 @registry.register_dataset("qm9")
