+++ ocpmodels/datasets/xie_grossman_mat_proj.py
@@ -1,5 +1,5 @@
-from baselines.common.registry import registry
-from baselines.datasets.base import BaseDataset
+from ocpmodels.common.registry import registry
+from ocpmodels.datasets.base import BaseDataset
 
 
 @registry.register_dataset("xie_grossman_mat_proj")
