+++ ocpmodels/models/cgcnn.py
@@ -2,13 +2,13 @@ import torch
 import torch.nn as nn
 from torch_geometric.nn import global_mean_pool
 
-from baselines.common.registry import registry
-from baselines.models.base import BaseModel
-from baselines.modules.layers import AttentionConv
+from ocpmodels.common.registry import registry
+from ocpmodels.models.base import BaseModel
+from ocpmodels.modules.layers import CGCNNConv
 
 
-@registry.register_model("transformer")
-class Transformer(BaseModel):
+@registry.register_model("cgcnn")
+class CGCNN(BaseModel):
     def __init__(
         self,
         num_atoms,
@@ -16,23 +16,16 @@ class Transformer(BaseModel):
         num_targets,
         atom_embedding_size=64,
         num_graph_conv_layers=6,
-        num_attention_heads=3,
-        attention_dropout=0,
         fc_feat_size=128,
         num_fc_layers=4,
     ):
-        super(Transformer, self).__init__(
-            num_atoms, bond_feat_dim, num_targets
-        )
+        super(CGCNN, self).__init__(num_atoms, bond_feat_dim, num_targets)
         self.embedding = nn.Linear(self.num_atoms, atom_embedding_size)
 
         self.convs = nn.ModuleList(
             [
-                AttentionConv(
-                    node_dim=atom_embedding_size,
-                    edge_dim=self.bond_feat_dim,
-                    num_heads=num_attention_heads,
-                    dropout=attention_dropout,
+                CGCNNConv(
+                    node_dim=atom_embedding_size, edge_dim=self.bond_feat_dim
                 )
                 for _ in range(num_graph_conv_layers)
             ]
@@ -51,12 +44,20 @@ class Transformer(BaseModel):
         self.fc_out = nn.Linear(fc_feat_size, self.num_targets)
 
     def forward(self, data):
-        node_feats = self.embedding(data.x)
-        for f in self.convs:
-            node_feats = f(node_feats, data.edge_index, data.edge_attr)
-        mol_feats = global_mean_pool(node_feats, data.batch)
+        mol_feats = self._convolve(data)
         mol_feats = self.conv_to_fc(mol_feats)
         if hasattr(self, "fcs"):
             mol_feats = self.fcs(mol_feats)
         out = self.fc_out(mol_feats)
         return out
+
+    def _convolve(self, data):
+        """
+        Returns the output of the convolution layers before they are passed
+        into the dense layers.
+        """
+        node_feats = self.embedding(data.x)
+        for f in self.convs:
+            node_feats = f(node_feats, data.edge_index, data.edge_attr)
+        mol_feats = global_mean_pool(node_feats, data.batch)
+        return mol_feats
