+++ ocpmodels/models/transformer.py
@@ -2,13 +2,13 @@ import torch
 import torch.nn as nn
 from torch_geometric.nn import global_mean_pool
 
-from baselines.common.registry import registry
-from baselines.models.base import BaseModel
-from baselines.modules.layers import CGCNNConv
+from ocpmodels.common.registry import registry
+from ocpmodels.models.base import BaseModel
+from ocpmodels.modules.layers import AttentionConv
 
 
-@registry.register_model("cgcnn")
-class CGCNN(BaseModel):
+@registry.register_model("transformer")
+class Transformer(BaseModel):
     def __init__(
         self,
         num_atoms,
@@ -16,16 +16,23 @@ class CGCNN(BaseModel):
         num_targets,
         atom_embedding_size=64,
         num_graph_conv_layers=6,
+        num_attention_heads=3,
+        attention_dropout=0,
         fc_feat_size=128,
         num_fc_layers=4,
     ):
-        super(CGCNN, self).__init__(num_atoms, bond_feat_dim, num_targets)
+        super(Transformer, self).__init__(
+            num_atoms, bond_feat_dim, num_targets
+        )
         self.embedding = nn.Linear(self.num_atoms, atom_embedding_size)
 
         self.convs = nn.ModuleList(
             [
-                CGCNNConv(
-                    node_dim=atom_embedding_size, edge_dim=self.bond_feat_dim
+                AttentionConv(
+                    node_dim=atom_embedding_size,
+                    edge_dim=self.bond_feat_dim,
+                    num_heads=num_attention_heads,
+                    dropout=attention_dropout,
                 )
                 for _ in range(num_graph_conv_layers)
             ]
@@ -44,20 +51,12 @@ class CGCNN(BaseModel):
         self.fc_out = nn.Linear(fc_feat_size, self.num_targets)
 
     def forward(self, data):
-        mol_feats = self._convolve(data)
+        node_feats = self.embedding(data.x)
+        for f in self.convs:
+            node_feats = f(node_feats, data.edge_index, data.edge_attr)
+        mol_feats = global_mean_pool(node_feats, data.batch)
         mol_feats = self.conv_to_fc(mol_feats)
         if hasattr(self, "fcs"):
             mol_feats = self.fcs(mol_feats)
         out = self.fc_out(mol_feats)
         return out
-
-    def _convolve(self, data):
-        """
-        Returns the output of the convolution layers before they are passed
-        into the dense layers.
-        """
-        node_feats = self.embedding(data.x)
-        for f in self.convs:
-            node_feats = f(node_feats, data.edge_index, data.edge_attr)
-        mol_feats = global_mean_pool(node_feats, data.batch)
-        return mol_feats
